% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirGroup.R
\docType{class}
\name{milk_mir_group_factory}
\alias{milk_mir_group_factory}
\title{Generate a MilkMirGroup object}
\format{An \code{\link{R6Class}} generator object}
\usage{
milk_mir_group_factory
}
\description{
Generate a MilkMirGroup object
}
\keyword{datasets}
