% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirSpectra.R
\docType{class}
\name{milk_mir_spectra_factory}
\alias{milk_mir_spectra_factory}
\title{Generate a MilkMirSpectra object}
\format{An \code{\link{R6Class}} generator object}
\usage{
milk_mir_spectra_factory
}
\description{
Generate a MilkMirSpectra object
}
\keyword{datasets}
