% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkMiR.R
\docType{package}
\name{milkMiR}
\alias{milkMiR}
\alias{milkMiR-package}
\title{Management of milk mid-infrared (MIR) spectra data}
\description{
milkMiR is a package to manage milk mid-infrared (MIR) spectra data
}
\author{
Tianfu Yang \email{tianfu@ualberta.ca}
Zhiquan Wang \email{zhiquan@ualberta.ca}
}

